initial drawings for the turn coordinator sensor
The plan is to make a diy "frictionless" rotational sensor. This disc
will be attached to a freely rotating rod. The rod will go upward out of
the top of the fuse (or wing maybe) and have a stiff flag attached to it
(acting like a fin or rudder). This will get held into position by the
oncoming airflow and rotate to either side to sense coordination. Much
like a yaw string on a glider. This disc will rotate above an analog
line following sensor that should be able to detect which "wedge" of the
disc it's looking at, and tell me the position of the flag. Might be
pretty nasty data, but probably worth a shot. Could be cool, if I can
get rough position at the cost of only one ADC on my micro.